Add files via upload
This project is done by using Python and OpenCV to implement 'Optical character recognition' from scratch, OCR is a technology used to extract text from images, such as photos or documents. In this project, I've implemented 'CCL' (Connected component labeling) to give each character a unique label and measured the bounding box of each character. Using SIFT got the features of all the characters from the test image and matched them with the features of the given characters, the 'SSD' algorithm was used for the feature matching and the corresponding character was assigned to the match with minimum error.